added invoice table and form
